Implement C++11 variadic templates based function_equal.
