+++ include/boost/phoenix/core/detail/preprocessed/function_equal.hpp
@@ -9,15 +9,15 @@
 #define BOOST_PHOENIX_PREPROCESSED_CORE_FUNCTION_EQUAL_HPP
 
 #if BOOST_PHOENIX_LIMIT <= 10
-#include <boost/phoenix/core/preprocessed/function_equal_10.hpp>
+#include <boost/phoenix/core/detail/preprocessed/function_equal_10.hpp>
 #elif BOOST_PHOENIX_LIMIT <= 20
-#include <boost/phoenix/core/preprocessed/function_equal_20.hpp>
+#include <boost/phoenix/core/detail/preprocessed/function_equal_20.hpp>
 #elif BOOST_PHOENIX_LIMIT <= 30
-#include <boost/phoenix/core/preprocessed/function_equal_30.hpp>
+#include <boost/phoenix/core/detail/preprocessed/function_equal_30.hpp>
 #elif BOOST_PHOENIX_LIMIT <= 40
-#include <boost/phoenix/core/preprocessed/function_equal_40.hpp>
+#include <boost/phoenix/core/detail/preprocessed/function_equal_40.hpp>
 #elif BOOST_PHOENIX_LIMIT <= 50
-#include <boost/phoenix/core/preprocessed/function_equal_50.hpp>
+#include <boost/phoenix/core/detail/preprocessed/function_equal_50.hpp>
 #else
 #error "BOOST_PHOENIX_LIMIT out of bounds for preprocessed headers"
 #endif
